+++ what4/src/What4/Expr/GroundEval.hs
@@ -19,6 +19,7 @@
 {-# LANGUAGE PolyKinds #-}
 {-# LANGUAGE RankNTypes #-}
 {-# LANGUAGE ScopedTypeVariables #-}
+{-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE TypeOperators #-}
 {-# LANGUAGE UndecidableInstances #-}
@@ -52,6 +53,9 @@ import qualified Data.Map.Strict as Map
 import           Data.Maybe ( fromMaybe )
 import           Data.Parameterized.Ctx
 import qualified Data.Parameterized.Context as Ctx
+import           Data.Parameterized.Fin
+                   ( Fin, addFinModN, minFin, mkFin, mulFinModN, negFinModN
+                   , recipFinModN )
 import           Data.Parameterized.NatRepr
 import           Data.Parameterized.TraversableFC
 import           Data.Ratio
@@ -85,6 +89,7 @@ type family GroundValue (tp :: BaseType) where
   GroundValue (BaseStringType si)   = StringLiteral si
   GroundValue (BaseArrayType idx b) = GroundArray idx b
   GroundValue (BaseStructType ctx)  = Ctx.Assignment GroundValueWrapper ctx
+  GroundValue (BaseFFType p)        = Fin p
 
 -- | Return a ground representation of a value, if it is ground.
 --
@@ -104,6 +109,7 @@ asGround x =
       def <- asConstantArray x
       groundDef <- asGround def
       pure (ArrayConcrete groundDef Map.empty)
+    BaseFFRepr{} -> asFF x
 
 -- | Inject a 'GroundValue' back into a 'SymExpr'.
 --
@@ -139,6 +145,7 @@ groundToSym sym tpr val =
            i' <- traverseFC (indexLit sym) i
            x' <- groundToSym sym tpr' x
            arrayUpdate sym arr' i' x'
+    BaseFFRepr p -> ffLit sym p val
 
 -- | A function that calculates ground values for elements.
 --   Clients of solvers should use the @groundEval@ function for computing
@@ -222,6 +229,8 @@ defaultValueForType tp =
     BaseArrayRepr _ b -> ArrayConcrete (defaultValueForType b) Map.empty
     BaseStructRepr ctx -> fmapFC (GVW . defaultValueForType) ctx
     BaseFloatRepr _fpp -> BF.bfPosZero
+    BaseFFRepr (_ :: NatRepr p)
+      | LeqProof <- leqSub (LeqProof @2 @p) (LeqProof @1 @2) -> minFin
 
 {-# INLINABLE evalGroundExpr #-}
 -- | Helper function for evaluating @Expr@ expressions in a model.
@@ -259,6 +268,7 @@ tryEvalGroundExpr :: (forall u . Expr t u -> MaybeT IO (GroundValue u))
 tryEvalGroundExpr _ (SemiRingLiteral SR.SemiRingIntegerRepr c _) = return c
 tryEvalGroundExpr _ (SemiRingLiteral SR.SemiRingRealRepr c _) = return c
 tryEvalGroundExpr _ (SemiRingLiteral (SR.SemiRingBVRepr _ _ ) c _) = return c
+tryEvalGroundExpr _ (SemiRingLiteral (SR.SemiRingFFRepr _) c _) = return c
 tryEvalGroundExpr _ (StringExpr x _)  = return x
 tryEvalGroundExpr _ (BoolExpr b _)    = return b
 tryEvalGroundExpr _ (FloatExpr _ f _) = return f
@@ -321,6 +331,7 @@ groundEq bt0 x0 y0 = unMAnd (f bt0 x0 y0)
         coerceMAnd (Ctx.traverseWithIndex
           (\i tp -> f tp (unGVW (x Ctx.! i)) (unGVW (y Ctx.! i))) flds)
       BaseArrayRepr{} -> MAnd Nothing
+      BaseFFRepr{} -> mand $ x == y
 
 -- | Helper function for evaluating @App@ expressions.
 --
@@ -384,6 +395,15 @@ evalGroundApp f a0 = do
            where
            smul sm e = BV.and sm <$> f e
            sadd x y  = pure (BV.xor x y)
+        SR.SemiRingFFRepr (p :: NatRepr p)
+          |  LeqProof <- leqSub (LeqProof @2 @p) (LeqProof @1 @2)
+          -> WSum.evalM sadd smul pure s
+           where
+           smul :: (1 <= p) => Fin p -> Expr t (BaseFFType p) -> MaybeT IO (Fin p)
+           smul sm e = mulFinModN p sm <$> f e
+
+           sadd :: (1 <= p) => Fin p -> Fin p -> MaybeT IO (Fin p)
+           sadd x y = pure (addFinModN p x y)
 
     SemiRingProd pd ->
       case WSum.prodRepr pd of
@@ -393,6 +413,9 @@ evalGroundApp f a0 = do
           fromMaybe (BV.one w) <$> WSum.prodEvalM (\x y -> pure (BV.mul w x y)) f pd
         SR.SemiRingBVRepr SR.BVBitsRepr w ->
           fromMaybe (BV.maxUnsigned w) <$> WSum.prodEvalM (\x y -> pure (BV.and x y)) f pd
+        SR.SemiRingFFRepr (p :: NatRepr p)
+          |  LeqProof <- leqSub (LeqProof @2 @p) (LeqProof @1 @2)
+          -> fromMaybe (mkFin (knownNat @1)) <$> WSum.prodEvalM (\x y -> pure (addFinModN p x y)) f pd
 
     RealDiv x y -> do
       xv <- f x
@@ -618,6 +641,17 @@ evalGroundApp f a0 = do
         (zip
           (BV.enumFromToUnsigned ground_lhs_idx (BV.sub w (BV.add w ground_lhs_idx ground_len) (BV.mkBV w 1)))
           (BV.enumFromToUnsigned ground_rhs_idx (BV.sub w (BV.add w ground_rhs_idx ground_len) (BV.mkBV w 1))))
+
+    ------------------------------------------------------------------------
+    -- Finite Field Operations
+
+    FFNeg (p :: NatRepr p) x
+      |  LeqProof <- leqSub (LeqProof @2 @p) (LeqProof @1 @2)
+      -> negFinModN p <$> f x
+    FFRecip (p :: NatRepr p) x
+      |  LeqProof <- leqSub (LeqProof @2 @p) (LeqProof @1 @2)
+      -> do r <- recipFinModN p <$> f x
+            MaybeT (pure r)
 
     ------------------------------------------------------------------------
     -- Conversions