+++ what4/CHANGES.md
@@ -33,6 +33,7 @@
   with a zero denominator.
 * Fix a bug in which calling `solver_adapter_write_smt2 yicesAdapter` would
   always fail.
+* TODO RGS: Mention finite field support
 
 # 1.7.3 -- 2026-01-26